refactor: 중복된 로딩 스피너 통합
